Refactored to emit less signals
Now releases less signals from deep within the beast and instead
releases a few from the model, but mostly a changed signal.

The view now listens for the changed signal and acts on that.
Controller returns the model in const and allows for the view
to directly request information about the model.

Signed-off-by: David Post <david@ohyonghao.com>